Move remaining files of tex2pdf_service
