+++ pyproject.toml
@@ -34,6 +34,7 @@ types-toml = "^0.10.8.20240310"
 click = "^8.1.7"
 tqdm = "^4.66.2"
 uvicorn = "^0.29.0"
+pre-commit = "*"
 
 
 [tool.ruff]
