+++ tex2pdf/service/tests/test_watermark.py
@@ -1,7 +1,7 @@
 import os
 import unittest
 
-from tex2pdf.pdf_watermark import add_watermark_text_to_pdf, Watermark
+from tex2pdf.service.pdf_watermark import add_watermark_text_to_pdf, Watermark
 
 watermark_pdf = "tests/test-output/watermark.pdf"
 in_pdf = "tests/fixture/smoke/Test.pdf"
@@ -13,7 +13,7 @@ class MyTestCase(unittest.TestCase):
 
     def test_watermarking(self):
         add_watermark_text_to_pdf(Watermark(
-                                    "Water World is in Orlando, FL.", 
+                                    "Water World is in Orlando, FL.",
                                     "https://en.wikipedia.org/wiki/Waterworld"
                                   ),
                                   in_pdf, "tests/test-output/Test.pdf"
