+++ tex2pdf/service/tex_patching.py
@@ -14,7 +14,7 @@ import os
 import re
 import typing
 
-from tex_inspection import TEX_FILE_EXTS
+from tex2pdf.tex_inspection import TEX_FILE_EXTS
 
 graphicspath_re = re.compile(r"\\graphicspath\{((\{.+?\})+)\}")
 paths_re = re.compile(r'\{(.+?)\}')
